7.16. Отрисуй меня полностью (часть 1) modified:   js/data.js modified:   js/popup.js modified:   js/util.js
